Move PlayerAppearance, PlayerColor, and PlayerPicker into player_settings module
